Check for the existence of stats_fn rather than dataset_dir
There was a bug where download_and_split_data would first create the dataset_dir and then error while actually making the stats; on subsequent runs, the code would try to access that nonexistent stats file since it was only checking for the existence of the dataset dir. This should fix that.